+++ src/config.rs
@@ -1,6 +1,5 @@
 use clap::Parser;
-use core::Language;
-use inspect_typescript::{Output, SortBy};
+use core::{Language, Output, SortBy};
 use serde::{Deserialize, Serialize};
 use std::env;
 use std::fs;
